refactor(tool): merge small single-purpose files into fewer files
