--- tool/cmd/otelc/cmd_go.go
@@ -1,19 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-package main
-
-import (
-	"github.com/urfave/cli/v3"
-	"go.opentelemetry.io/otelc/tool/internal/setup"
-)
-
-//nolint:gochecknoglobals // Implementation of a CLI command
-var commandGo = cli.Command{
-	Name:            "go",
-	Description:     "Invoke the go toolchain with toolexec mode",
-	ArgsUsage:       "[go toolchain flags]",
-	SkipFlagParsing: true,
-	Before:          addLoggerPhaseAttribute,
-	Action:          setup.GoBuild,
-}
