+++ tool/cmd/otelc/command_test.go
