--- tool/internal/instrument/apply_call_helpers.go
@@ -1,55 +1,0 @@
-// Copyright The OpenTelemetry Authors
-// SPDX-License-Identifier: Apache-2.0
-
-package instrument
-
-import (
-	"github.com/dave/dst"
-
-	"go.opentelemetry.io/otelc/tool/internal/rule"
-)
-
-// matchesCallRule checks if a call expression matches the rule's criteria.
-//
-// Only qualified calls are supported: pkg.Function()
-// The function_call rule must specify the full import path: "package/path.FunctionName"
-//
-// Examples in source code:
-//   - http.Get() after "import 'net/http'" matches "net/http.Get"
-//   - redis.Get() after "import redis 'github.com/redis/go-redis/v9'" matches "github.com/redis/go-redis/v9.Get"
-//   - sql.Open() after "import 'database/sql'" matches "database/sql.Open"
-//
-// What does NOT match:
-//   - Get() without package qualifier (unqualified calls not supported)
-//   - other.Get() where other is from a different package
-func matchesCallRule(call *dst.CallExpr, r *rule.InstCallRule, importAliases map[string]string) bool {
-	// Use pre-parsed fields - no parsing needed!
-	importPath := r.ImportPath
-	funcName := r.FuncName
-
-	// Only match qualified calls: pkg.Function()
-	sel, ok := call.Fun.(*dst.SelectorExpr)
-	if !ok {
-		return false
-	}
-
-	// Check function name matches
-	if sel.Sel.Name != funcName {
-		return false
-	}
-
-	// Check that the package identifier is a simple identifier (not a chained selector)
-	ident, ok := sel.X.(*dst.Ident)
-	if !ok {
-		return false
-	}
-
-	// Check that the package's import path matches the rule's import path.
-	pkgPath := ident.Path
-	if pkgPath != "" {
-		return pkgPath == importPath
-	}
-
-	resolvedPath, ok := importAliases[ident.Name]
-	return ok && resolvedPath == importPath
-}
